Move PaneSide to ui/mod.rs and relocate word-diff test to document.rs
